third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/max-length-exceeded-name.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  const name_too_long = 'a'.repeat(249);
  return assert_promise_rejects_with_message(
    requestDeviceWithKeyDown({filters: [{name: name_too_long}]}),
      new DOMException(
        'Failed to execute \'requestDevice\' on \'Bluetooth\': A device ' +
        'name can\'t be longer than 248 bytes.',
      'TypeError'),
    'Device name longer than 248');
}, 'A device name longer than 248 must reject.');
</script>
